Make standardised datasets

make_standardised_dss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make standardised datasets. The function returns Standardised datasets (a list).

Usage

make_standardised_dss(
  label_1L_chr = "Country",
  lookup_Ready4useDyad = ready4use::Ready4useDyad(),
  seed_Ready4useDyad = ready4use::Ready4useDyad(),
  standards_Ready4useDyad = ready4use::Ready4useDyad(),
  type_chr = c("Input", "Output"),
  validation_ls = list(),
  what_chr = c("Correspondences", "Lookup", "Seed", "Standards", "Validation"),
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  correspondences_y_r3 = ready4show::ready4show_correspondences()
)

Arguments

label_1L_chr

Label (a character vector of length one), Default: 'Country'

lookup_Ready4useDyad

PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()

seed_Ready4useDyad

PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()

standards_Ready4useDyad

PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()

type_chr

Type (a character vector), Default: c("Input", "Output")

validation_ls

Validation (a list), Default: list()

what_chr

What (a character vector), Default: c("Correspondences", "Lookup", "Seed", "Standards", "Validation")

correspondences_x_r3

Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()

correspondences_y_r3

Correspondences y (a ready4 S3), Default: ready4show::ready4show_correspondences()

Value

Standardised datasets (a list)