Update correspondences

update_correspondences() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update correspondences. Function argument correspondences_x_r3 specifies the object to be updated. Argument standards_df provides the object to be updated. The function returns Correspondences x (a ready4 S3).

Usage

update_correspondences(
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  standards_df,
  standards_var_nms_chr,
  seed_df,
  reference_var_nm_1L_chr,
  force_standard_1L_lgl = T,
  fuzzy_logic_1L_chr = "jw",
  max_distance_1L_dbl = Inf
)

Arguments

correspondences_x_r3

Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()

standards_df

Standards (a data.frame)

standards_var_nms_chr

Standards variable names (a character vector)

seed_df

Seed (a data.frame)

reference_var_nm_1L_chr

Reference variable name (a character vector of length one)

force_standard_1L_lgl

Force standard (a logical vector of length one), Default: T

fuzzy_logic_1L_chr

Fuzzy logic (a character vector of length one), Default: 'jw'

max_distance_1L_dbl

Maximum distance (a double vector of length one), Default: Inf

Value

Correspondences x (a ready4 S3)